requirements.txt
openai
psycopg2-binary
argparse
python-dotenv
autogen-agentchat
autogen-ext[openai]
